finished level

+++ MonsterMegaAdUnity/Assets/Scripts/StateMachine.cs
@@ -241,7 +241,7 @@ public class AttackState : IState
         StateMachine stateMachine = owner.GetComponent<StateMachine>();
         float range = owner.GetComponent<Enemy>().m_attackRange;
         float boxSize = owner.GetComponent<Enemy>().m_attackBoxRange;
-        float damage = owner.GetComponent<Enemy>().m_damage;
+        int damage = owner.GetComponent<Enemy>().m_damage;
         if (Vector2.Distance(owner.transform.position, target.transform.position) >= (stateMachine.m_attackRange)) stateMachine.ChangeState(new ChaseState(target, false));
         Collider2D hitCollider = Physics2D.OverlapBox(owner.transform.position, new Vector2(boxSize, boxSize), 0, LayerMask.GetMask("Entity"));
         if (hitCollider && hitCollider.CompareTag("Player"))
